Separate SM integrations from main CommandRouter through subclassing
